More consistent data spec for requirements
When specifying new software requirements in a put to a jobtype, the
required software and its min and max versions are specified as strings,
while a POST to the requirements collection for a jobtype would expect
those as ids.

This commit changes the latter endpoint to expect the required software
and its versions as strings as well.