Switch TaskLogs FKs to passive deletes
The former attempt at solving the problem did not work.

This one uses the underlying SQL database's ON DELETE CASCADE function
to achieve auto-deletion of task log associations on deletion of the
associated task or log, which means SQLAlchemy does not first have to
load the entire list of relationships before deleting an item.