Index suitable agents for jobtypes
This saves us a lot of manually checking whether a given agent will be
able to run a given job. Since these checks require one database query
per job, the resulting speed-up can be significant.

This is particularly obvious when there are some agents in the farm that
cannot run most kinds of jobs and will often sit idle while the rest of
the farm is busy.