Rollback, don't commit left over transactions
Celery keeps the state of global objects inside of the same worker from
one execution of a task to the next.  In this case, that mean db.session
objects are carried over, including their transaction state.

If there is a transaction left over from a previous task, it is better
to roll it back than to commit it.  This should also help with
db.session objects that are caught in an invalid transaction state,
because those not an explicit rollback to start a new transaction.