Rework delete_task celery task to be more robust
Checking whether the job this task belonged to should be deleted as well
is now done in a separate, new transaction.  This means that, at that
point, we also get to see changes that other transactions have made in
the time since our first one started.  This should help with some
concurrency issues that used to plague job deletions.

Failed transactions are now explicitly rolled back and retried up to ten
times.