Remove weight column from jobs list again
This is from user feedback. The reasoning here is that an individual
job's weight (as opposed to the weight of a job queue) will rarely if
ever be relevant, because the scheduler usually won't start more than
one job in a queue at a time, and if it does that's because the running
jobs cannot make use of all the available agents.