Use lock when sending notifies, track sent status
In spite of all the work that has gone into this in the past, our users
still keep seeing multiple notification emails for a single job.  This
patch should solve that for good.  It uses a combination of a database
field to track whether a notification has already been sent and a
filelock per job, so that even with low transaction isolation levels,
like "read committed", no races should occur.