Rework default software for agents
Software versions now have a "default" flag. If this is set, that
version will automatically be registered as available on all new agents.

This should be more user friendly and less prone to error than the
previous solution to this problem, since now a user cannot specify a
version that doesn't even exist as a default.