Remove login check for top level links
This way, all visitors, even those not logged in, will be able to the
top-level links to agents, jobs and so on.

Rationale: We do not have working authorization checks for the pages
anyway. If someone simply adds /jobs/ to the URL, they can view the jobs
list anyway. As it stands right now, this check is just a minor
annoyance at best and at worst will give users a false sense of
security.

We can readd this check once we do have real authorization.