Implement job groups
Job groups are a way to group jobs that form a logical unit (and may or
may not have an arbitrarily complex dependency hierarchy among them)
together under one convenient handle. The intention is mainly to use
them for a more concise display of jobs that doesn't have five entries
for something the user would consider to be just one (complex) job.