removing debug config and Python 3.3
With the latest changes the 'debug' configuration
no longer has a meaning.  For a while now it has
not really served much purpose so this is just
conslidating the existing code a bit.

This change also removes testing for Python 3.3 on
travis.  We initially had this to ensure that the
two latest 3.x releases worked.  However since adding
3.3 to the tests there has not been any discernible
differences between it and 3.4 either in our code or
in any of the packages we're using.