define a default for PYFARM_SECRET_KEY
It used to be that only in the 'debug' configuration
you'd get a default defined.  In most production situations
the secret key will be something an administrator will define.  In
other situations, such as the home user, it's not something someone
would think about.  This is a tradoff between security and ease of
use that we should revisit later (like when we have users/role auth again).