remove Python 3.3 from travis tests
Python 3.x and Python 2.x are our main targets
for testing PyFarm.  3.3 and 3.4 were both
originally included to ensure there were not
any differences in both the tests and the libraries
we're using.  So far there have not been measurable
differences between the two versions so we're removing
the extra version to save some time.  It can always
be added back later, especially if a library begins
to have issues with 3.x in general.