removing some of the scheduler environment variables
So because the config now allows timedeltas be specified
we can't accept an integer from an environment variable
without some extra code.  We can add the code necessary
to support this, but we'd only be removing it again a
short time later.  Another reason why this will probably
be fine for now, unless there are objections, is the default
values in the config mirror what the original environment
variables used.