Improve display of parent and child jobs
Gets rid of the state column in favor of an icon next to the title, as
has been done in many other places already.

Gets rid of the user and submitted columns, since they are too unlikely to
be different from the current job.

Adds a jobtype column, because that is actually interesting information
in that context.